Move tests into appropriate packages
